only generate route 'route-not-found' if router.js exits

+++ blueprints/@lblod/ember-vo-webuniversum/index.js
@@ -8,9 +8,16 @@ module.exports = {
   normalizeEntityName: function() {},
   async afterInstall(options) {
     const source = path.join(options.project.root, 'app', 'router.js');
-    const routes = new EmberRouterGenerator(fs.readFileSync(source));
-    var newRoutes = routes.add('route-not-found', {path: '/*wildcard'});
-    fs.writeFileSync(source, newRoutes.code());
+
+    if (fs.existsSync(source)) {
+      const routes = new EmberRouterGenerator(fs.readFileSync(source));
+      var newRoutes = routes.add('route-not-found', {path: '/*wildcard'});
+      fs.writeFileSync(source, newRoutes.code());
+    }
+    else{
+      console.log('No router.js found, skipping route-not-found');
+    }
+
     this.addPackageToProject('postscribe', '^2.0.8');
     //This will follow major versions until 4
     return this.addAddonToProject('ember-cli-moment-shim', '~3.7.1');
